# enterprise-ai-knowledge-assistant/Asset-Manager/Asset-Manager/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Enterprise AI Knowledge Assistant API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: documents
    description: Document management
  - name: conversations
    description: Chat conversations
  - name: dashboard
    description: Dashboard stats
  - name: admin
    description: Admin operations

security:
  - BearerAuth: []

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      security: []
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Email already registered

  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

  /documents/upload:
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUpload"
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/search:
    get:
      operationId: searchDocuments
      tags: [documents]
      summary: Semantic search across documents
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResult"

  /documents/stats:
    get:
      operationId: getDocumentStats
      tags: [documents]
      summary: Get document statistics
      responses:
        "200":
          description: Document statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentStats"

  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  /documents/{id}/chunks:
    get:
      operationId: getDocumentChunks
      tags: [documents]
      summary: Get document chunks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chunks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chunk"

  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      summary: List conversations
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [conversations]
      summary: Create a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  /conversations/{id}:
    get:
      operationId: getConversation
      tags: [conversations]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationDetail"
    delete:
      operationId: deleteConversation
      tags: [conversations]
      summary: Delete conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  /conversations/{id}/messages:
    post:
      operationId: sendMessage
      tags: [conversations]
      summary: Send a message and get AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: AI response with citations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardData"

  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"

  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Delete user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  /admin/documents:
    get:
      operationId: listAdminDocuments
      tags: [admin]
      summary: List all documents (admin)
      responses:
        "200":
          description: All documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminDocument"

  /admin/documents/{id}:
    delete:
      operationId: deleteAdminDocument
      tags: [admin]
      summary: Delete any document (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
      required: [id, email, full_name, role, is_active, created_at]

    UserRegistration:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
        full_name:
          type: string
      required: [email, password, full_name]

    UserLogin:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]

    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [access_token, token_type, user]

    UserUpdate:
      type: object
      properties:
        role:
          type: ["string", "null"]
        is_active:
          type: ["boolean", "null"]

    Document:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        file_type:
          type: string
        size_bytes:
          type: integer
        status:
          type: string
        category:
          type: ["string", "null"]
        topics:
          type: array
          items:
            type: string
        page_count:
          type: integer
        chunk_count:
          type: integer
        created_at:
          type: string
          format: date-time
        uploaded_by_id:
          type: integer
      required: [id, filename, file_type, size_bytes, status, page_count, chunk_count, created_at, uploaded_by_id]

    DocumentUpload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: Document file (PDF, DOCX, TXT, CSV)

    DocumentStats:
      type: object
      properties:
        total_documents:
          type: integer
        total_chunks:
          type: integer
        categories:
          type: object
          additionalProperties:
            type: integer
        document_types:
          type: object
          additionalProperties:
            type: integer
        processing_count:
          type: integer
        ready_count:
          type: integer
        error_count:
          type: integer
      required: [total_documents, total_chunks, categories, document_types, processing_count, ready_count, error_count]

    SearchResult:
      type: object
      properties:
        document_id:
          type: integer
        filename:
          type: string
        chunk_text:
          type: string
        score:
          type: number
        page_number:
          type: integer
      required: [document_id, filename, chunk_text, score, page_number]

    Chunk:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        page_number:
          type: integer
        chunk_index:
          type: integer
      required: [id, text, page_number, chunk_index]

    Conversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: ["string", "null"]
          format: date-time
        message_count:
          type: integer
      required: [id, title, created_at, message_count]

    ConversationInput:
      type: object
      properties:
        title:
          type: string
          default: "New Conversation"
      required: [title]

    ConversationDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: ["string", "null"]
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
      required: [id, title, created_at, messages]

    Message:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
        content:
          type: string
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
        created_at:
          type: string
          format: date-time
      required: [id, role, content, citations, created_at]

    MessageInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
      required: [content]

    Citation:
      type: object
      properties:
        document_id:
          type: integer
        filename:
          type: string
        chunk_text:
          type: string
        page_number:
          type: integer
        score:
          type: number
      required: [document_id, filename, chunk_text, page_number, score]

    ChatResponse:
      type: object
      properties:
        message_id:
          type: integer
        content:
          type: string
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
        model:
          type: ["string", "null"]
      required: [message_id, content, citations]

    DashboardData:
      type: object
      properties:
        total_documents:
          type: integer
        total_conversations:
          type: integer
        total_messages:
          type: integer
        recent_documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        recent_conversations:
          type: array
          items:
            $ref: "#/components/schemas/Conversation"
        documents_by_category:
          type: object
          additionalProperties:
            type: integer
        documents_by_type:
          type: object
          additionalProperties:
            type: integer
      required: [total_documents, total_conversations, total_messages, recent_documents, recent_conversations, documents_by_category, documents_by_type]

    AdminUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        document_count:
          type: integer
        conversation_count:
          type: integer
      required: [id, email, full_name, role, is_active, created_at, document_count, conversation_count]

    AdminDocument:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        file_type:
          type: string
        size_bytes:
          type: integer
        status:
          type: string
        category:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        uploaded_by_email:
          type: string
      required: [id, filename, file_type, size_bytes, status, created_at, uploaded_by_email]

    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]