# enterprise-ai-knowledge-assistant/Asset-Manager/Asset-Manager/lib/api-zod/src/generated/types/citation.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise AI Knowledge Assistant API
 * OpenAPI spec version: 1.0.0
 */

export interface Citation {
  document_id: number;
  filename: string;
  chunk_text: string;
  page_number: number;
  score: number;
}