# enterprise-ai-knowledge-assistant/Asset-Manager/Asset-Manager/lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise AI Knowledge Assistant API
 * OpenAPI spec version: 1.0.0
 */
import type { Citation } from './citation';

export interface Message {
  id: number;
  role: string;
  content: string;
  citations: Citation[];
  created_at: Date;
}